chore: convert more code to Kotlin
